Remove runtime type checking
Our APIs generally prefer leaving out type checking, not least to avoid
overhead, but also because the contract should be clearly defined and
callers assumed to adhere to the contract. For more formal checking of
the contract, TypeScript is preferred and we should also provide type
declaration files for this module.